stable version before plotting
