+++ archived_implementation/a_pulse_histograms.py
